+++ src/ralph/defaults.py
@@ -6,7 +6,14 @@ from pathlib import Path
 
 from click import get_app_dir
 
-from .utils import import_string
+
+class DatabaseBackends(Enum):
+    """Enumerate active database backend modules.
+
+    Adding an entry to this enum will make it available to the CLI.
+    """
+
+    ES = "ralph.backends.database.es.ESDatabase"
 
 
 class Parsers(Enum):
@@ -29,11 +36,8 @@ class StorageBackends(Enum):
 
 
 APP_DIR = Path(environ.get("RALPH_APP_DIR", get_app_dir("ralph")))
-AVAILABLE_PARSERS = (lambda: (import_string(parser.value).name for parser in Parsers))()
-AVAILABLE_STORAGE_BACKENDS = (
-    lambda: (import_string(backend.value).name for backend in StorageBackends)
-)()
 DEFAULT_GELF_PARSER_CHUNCK_SIZE = 5000
+DEFAULT_BACKEND_CHUNCK_SIZE = 500
 ENVVAR_PREFIX = "RALPH"
 HISTORY_FILE = Path(environ.get("RALPH_HISTORY_FILE", APP_DIR / "history.json"))
 FS_STORAGE_DEFAULT_PATH = Path("./archives")