+++ src/ralph/exceptions.py
@@ -9,3 +9,7 @@ class BackendParameterException(Exception):
 
 class EventKeyError(Exception):
     """Raised when an expected event key has not been found."""
+
+
+class UnsupportedBackendException(Exception):
+    """Raised when trying to use an unsupported backend type"""